+++ eoreader/env_vars.py
@@ -115,3 +115,9 @@ LEGACY_BAND_NAME_RESOLUTION = "EOREADER_LEGACY_BAND_NAME_RESOLUTION"
 """
 Keep legacy resolution in band name (:code:`1000-00m` instead of :code:`1000m`, or :code:`0-50m` instead of :code:`0-5m`)
 """
+
+FIX_MAXAR = "EOREADER_FIX_MAXAR"
+"""
+Fix faulty Maxar product (corrupted shapes in metadata). 
+This requires an alteration of the raw data, hence the possibility to block it by setting this environment variable to 0.
+"""